% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\name{coef.infometrics_me}
\alias{coef.infometrics_me}
\title{Extract estimated probabilities from an infometrics_me object}
\usage{
\method{coef}{infometrics_me}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"infometrics_me"}.}

\item{...}{Not used.}
}
\value{
Named numeric vector of estimated probabilities.
}
\description{
Extract estimated probabilities from an infometrics_me object
}
