% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{.check_prob_pair}
\alias{.check_prob_pair}
\title{Validate a pair of compatible probability vectors}
\usage{
.check_prob_pair(p, q)
}
\description{
Validate a pair of compatible probability vectors
}
\keyword{internal}
