% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinomial_gce.R
\name{fano_bounds}
\alias{fano_bounds}
\title{Information-Theoretic (Fano) Error Bounds}
\usage{
fano_bounds(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Passed to methods.}
}
\value{
A method-specific object of error bounds.
}
\description{
Generic for Fano's-inequality error bounds on a fitted probability matrix.
}
\seealso{
\code{\link{fano_bounds.multinomial_gce}}
}
