% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\name{fitted.infometrics_me}
\alias{fitted.infometrics_me}
\title{Fitted (moment) values from an infometrics_me object}
\usage{
\method{fitted}{infometrics_me}(object, ...)
}
\arguments{
\item{object}{An object of class \code{"infometrics_me"}.}

\item{...}{Not used.}
}
\value{
Numeric vector of fitted moment values X %*% p_hat.
}
\description{
Fitted (moment) values from an infometrics_me object
}
