% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\name{print.infometrics_me}
\alias{print.infometrics_me}
\title{Print method for infometrics_me objects}
\usage{
\method{print}{infometrics_me}(x, digits = 4, ...)
}
\arguments{
\item{x}{An object of class \code{"infometrics_me"}.}

\item{digits}{Integer. Number of significant digits (default 4).}

\item{...}{Not used.}
}
\description{
Print method for infometrics_me objects
}
