% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me.R
\name{summary.infometrics_me}
\alias{summary.infometrics_me}
\title{Summary method for infometrics_me objects}
\usage{
\method{summary}{infometrics_me}(object, digits = 4, ...)
}
\arguments{
\item{object}{An object of class \code{"infometrics_me"}.}

\item{digits}{Integer. Number of significant digits (default 4).}

\item{...}{Not used.}
}
\description{
Summary method for infometrics_me objects
}
